Hot 热更错误处理第一版。

+++ ZezeJava/ZezeJava/src/main/java/Zeze/Hot/HotManager.java
@@ -24,6 +24,7 @@ import Zeze.IModule;
 import Zeze.Schemas;
 import Zeze.Serialize.ByteBuffer;
 import Zeze.Transaction.Bean;
+import Zeze.Util.Action0;
 import Zeze.Util.ConcurrentHashSet;
 import Zeze.Util.FewModifyMap;
 import Zeze.Util.FewModifySortedMap;
@@ -231,7 +232,6 @@ public class HotManager extends ClassLoader {
 		}
 	}
 
-	// todo【警告】安装过程目前没有事务性，如果安装出现错误，可能导致某些模块出错。
 	private boolean upgrading = false;
 
 	public boolean isUpgrading() {
@@ -264,24 +264,58 @@ public class HotManager extends ClassLoader {
 			modules.put(exist.getName(), exist);
 		}
 		// 发生错误的禁止暴露服务。
-		exists.get(reverseErrorIndex).disable();
+		if (reverseErrorIndex >= 0)
+			exists.get(reverseErrorIndex).disable();
 	}
 
-	private List<HotModule> install(List<String> namespaces) throws Exception {
+	public class MainRollbackAction implements Action0 {
+		private final Zeze.Application zeze;
+		private final ArrayList<HotModule> exists;
+		private final Schemas oldSchemas;
+		private boolean hasAlter = false;
+
+		public MainRollbackAction(Zeze.Application zeze, ArrayList<HotModule> exists, Schemas oldSchemas) {
+			this.zeze = zeze;
+			this.exists = exists;
+			this.oldSchemas = oldSchemas;
+		}
+
+		public void setHasAlter() {
+			hasAlter = true;
+		}
+
+		@Override
+		public void run() throws Exception {
+			// 恢复schemas
+			zeze.__upgrade_schemas__(oldSchemas);
+			// 导致HotModule重新start并导致Register。
+			recoverModules(exists, -1);
+			// 在恢复的Schemas的旧结构下，重新alter。
+			if (hasAlter)
+				zeze.__install_alter__();
+			// 内存表的回滚保持旧数据即可，在TableX.open(exist, app)里面处理。
+			zeze.__get_upgrade_memory_table__().clear();
+		}
+	}
+
+	private ArrayList<HotModule> install(List<String> namespaces) throws Exception {
 		logger.info("________________ install ________________ {}", namespaces);
+		// 首先查询得到现有的模块，不存在也加入null。
+		var exists = new ArrayList<HotModule>(namespaces.size());
+		for (var namespace : namespaces)
+			exists.add(modules.get(namespace));
+		// 锁外执行stopBefore
+		for (var exist : exists)
+			exist.stopBefore();
+		var result = new ArrayList<HotModule>();
 		try (var ignored = enterWriteLock()) {
 			upgrading = true;
 			var app = zeze.getAppBase();
-			app.getZeze().__install_prepare__(loadSchemas());
-
-			var result = new ArrayList<HotModule>(namespaces.size());
-			var exists = new ArrayList<HotModule>(namespaces.size());
-
 			// 先保存现有的数据
 			app.getZeze().checkpointRun();
 			// remove
 			for (var namespace : namespaces) {
-				exists.add(modules.remove(namespace));
+				modules.remove(namespace);
 			}
 			// reverse stop
 			var reverseIndex = exists.size() - 1;
@@ -293,83 +327,125 @@ public class HotManager extends ClassLoader {
 				}
 			} catch (Throwable ex) {
 				logger.error("stop modules {}", exists, ex);
-				recoverModules(exists, reverseIndex);
-				return result; // isEmpty
+				recoverModules(exists, reverseIndex); // 此时的恢复，schemas没有变化，不需要重新alter。
+				return null;
 			}
-			var freshHotUpgrades = new ArrayList<HotUpgrade>();
-			for (var hotUpgrade : hotUpgrades) {
-				if (hotUpgrade.hasFreshStopModuleLocalOnce()) {
-					freshHotUpgrades.add(hotUpgrade);
-				}
-			}
-			var freshHotBeanFactories = new ArrayList<HotBeanFactory>();
-			for (var hotBeanFactory : hotBeanFactories) {
-				if (hotBeanFactory.hasFreshStopModuleDynamicOnce()) {
-					freshHotBeanFactories.add(hotBeanFactory);
-				}
-			}
-			// install
-			// todo 这里都是内部代码，本来不容易出错。但是涉及文件操作，有可能发生错误，而且非常难处理。
-			//  感觉上应该停止整个程序，重启比较可靠。
-			for (var namespace : namespaces) {
-				result.add(_install(namespace));
-			}
-			// batch load redirect
-			var iModules = createModuleInstance(result);
-			app.getZeze().__install_alter__();
 
-			var hotJarFiles = new ArrayList<JarFile>();
-			var newModules = new ArrayList<HotModule>();
-			for (var ii = 0; ii < iModules.length; ++ii) {
-				var exist = exists.get(ii);
-				var module = result.get(ii);
-				hotJarFiles.add(module.getJarFile());
-				module.setService(iModules[ii]);
-				if (null != exist) {
-					try {
-						app.getZeze().newProcedure(() -> {
+			var txn = new HotTransaction("HotManager.install");
+			app.getZeze().__install_prepare__();
+			var old = app.getZeze().__upgrade_schemas__(loadSchemas());
+
+			var mainRollbackAction = new MainRollbackAction(zeze, exists, old);
+			txn.whileRollback(mainRollbackAction);
+
+			ArrayList<HotUpgrade> freshHotUpgrades;
+			ArrayList<HotBeanFactory> freshHotBeanFactories;
+			ArrayList<JarFile> hotJarFiles;
+			ArrayList<HotModule> newModules;
+			try {
+				freshHotUpgrades = new ArrayList<>();
+				for (var hotUpgrade : hotUpgrades) {
+					if (hotUpgrade.hasFreshStopModuleLocalOnce()) {
+						freshHotUpgrades.add(hotUpgrade);
+					}
+				}
+				freshHotBeanFactories = new ArrayList<>();
+				for (var hotBeanFactory : hotBeanFactories) {
+					if (hotBeanFactory.hasFreshStopModuleDynamicOnce()) {
+						freshHotBeanFactories.add(hotBeanFactory);
+					}
+				}
+				// install
+				for (var namespace : namespaces) {
+					result.add(_install(namespace, txn));
+				}
+				// batch load redirect
+				var iModules = createModuleInstance(result);
+				app.getZeze().__install_alter__();
+				mainRollbackAction.setHasAlter();
+
+				hotJarFiles = new ArrayList<>();
+				newModules = new ArrayList<>();
+				for (var ii = 0; ii < iModules.length; ++ii) {
+					var exist = exists.get(ii);
+					var module = result.get(ii);
+					hotJarFiles.add(module.getJarFile());
+					module.setService(iModules[ii]);
+					if (null != exist) {
+						var rc = app.getZeze().newProcedure(() -> {
 							module.upgrade(exist);
 							return 0;
 						}, "HotManager.upgrade").call();
-					} catch (Throwable anyError) {
-						logger.error("todo some rollback need.", anyError);
+						if (0 != rc)
+							throw new RuntimeException("upgrade " + module + " rc=" + rc);
+					} else {
+						newModules.add(module);
 					}
-				} else {
-					newModules.add(module);
 				}
+				txn.commit();
+			} catch (Throwable ex) {
+				logger.error("", ex);
+				txn.rollback();
+				return null;
 			}
-			// internal upgrade
-			for (var hotUpgrade : freshHotUpgrades) {
-				hotUpgrade.upgrade((bean) -> retreat(exists, result, bean));
+			// 下面进入”不能“出错阶段：1. start 出错则相关模块停止服务；2. 其他错误，停止程序。
+			try {
+				// internal upgrade
+				for (var hotUpgrade : freshHotUpgrades) {
+					hotUpgrade.upgrade((bean) -> retreat(exists, result, bean));
+				}
+				// upgrade user memory table
+				for (var table : app.getZeze().__get_upgrade_memory_table__()) {
+					table.upgrade();
+				}
+				app.getZeze().__get_upgrade_memory_table__().clear();
+				var beanFactories = new HashMap<BeanFactory, List<Class<?>>>();
+				for (var hotBeanFactory : freshHotBeanFactories) {
+					hotBeanFactory.clearTableCache();
+					beanFactories.put(hotBeanFactory.beanFactory(), new ArrayList<>());
+				}
+				BeanFactory.resetHot(beanFactories, hotJarFiles);
+				for (var hotBeanFactory : freshHotBeanFactories) {
+					hotBeanFactory.processWithNewClasses(beanFactories.get(hotBeanFactory.beanFactory()));
+				}
+				// start ordered
+				var startErrors = new ArrayList<HotModule>();
+				for (var module : result) {
+					try {
+						module.start();
+					} catch (Exception ex) {
+						logger.error("", ex);
+						startErrors.add(module);
+					}
+				}
+				// 再次停止启动失败的模块，并注销。
+				for (var startErrorIndex = startErrors.size() - 1; startErrorIndex >= 0; --startErrorIndex) {
+					var module = modules.remove(startErrors.get(startErrorIndex).getName());
+					module.stop();
+					module.disable();
+				}
+				for (var module : newModules) {
+					if (startErrors.contains(module))
+						continue;
+					var moduleConfig = module.loadModuleConfig();
+					zeze.getProviderApp().providerService.addHotModule((IModule)module.getService(), moduleConfig);
+				}
+			} catch (Throwable ex) {
+				logger.error("too much changes. impossible to rollback.", ex);
+				LogManager.shutdown();
+				Runtime.getRuntime().halt(111222);
 			}
-			// upgrade user memory table
-			for (var table : app.getZeze().__get_upgrade_memory_table__()) {
-				table.upgrade();
-			}
-			app.getZeze().__get_upgrade_memory_table__().clear();
-			var beanFactories = new HashMap<BeanFactory, List<Class<?>>>();
-			for (var hotBeanFactory : freshHotBeanFactories) {
-				hotBeanFactory.clearTableCache();
-				beanFactories.put(hotBeanFactory.beanFactory(), new ArrayList<>());
-			}
-			BeanFactory.resetHot(beanFactories, hotJarFiles);
-			for (var hotBeanFactory : freshHotBeanFactories) {
-				hotBeanFactory.processWithNewClasses(beanFactories.get(hotBeanFactory.beanFactory()));
-			}
-
-			// start ordered todo 错误处理
-			for (var module : result) {
-				module.start();
-			}
-
-			for (var module : newModules) {
-				var moduleConfig = module.loadModuleConfig();
-				zeze.getProviderApp().providerService.addHotModule((IModule)module.getService(), moduleConfig);
-			}
-			return result;
 		} finally {
 			upgrading = false;
 		}
+		for (var module : result) {
+			try {
+				module.startLast();
+			} catch (Exception ex) {
+				logger.error("", ex);
+			}
+		}
+		return result;
 	}
 
 	private void putJar(File file) throws IOException {
@@ -391,7 +467,7 @@ public class HotManager extends ClassLoader {
 	 * @return HotModule
 	 * @throws Exception error
 	 */
-	private HotModule _install(String namespace) throws Exception {
+	private HotModule _install(String namespace, HotTransaction txn) throws Exception {
 		var moduleSrc = Path.of(distributeDir, namespace + ".jar").toFile();
 		var interfaceSrc = Path.of(distributeDir, namespace + ".interface.jar").toFile();
 		if (!moduleSrc.exists() || !interfaceSrc.exists())
@@ -404,14 +480,15 @@ public class HotManager extends ClassLoader {
 			if (null != oldI)
 				oldI.close();
 		}
-		for (var i = 0; i < 10; ++i) {
-			try {
-				if (Files.deleteIfExists(interfaceDst.toPath()))
-					break;
-			} catch (Exception ignored) {
-			}
-			//System.out.println("delete interface fail=" + i + ":" + interfaceDstAbsolute);
-			Thread.sleep(100);
+		var interfaceDstBackup = Path.of(workingDir, "interfaces", namespace + ".interface.jar.backup").toFile();
+		if (interfaceDst.renameTo(interfaceDstBackup)) {
+			// 从备份恢复，并且重新加载旧文件。
+			txn.whileRollback(() -> {
+				if (interfaceDstBackup.renameTo(interfaceDst))
+					putJar(interfaceDst);
+			});
+			// 提交的时候删除备份
+			txn.whileCommit(() -> Files.deleteIfExists(interfaceDstBackup.toPath()));
 		}
 		if (!interfaceSrc.renameTo(interfaceDst))
 			throw new RuntimeException("rename fail. " + interfaceSrc + "->" + interfaceDst);
@@ -419,14 +496,13 @@ public class HotManager extends ClassLoader {
 
 		// 安装 module
 		var moduleDst = Path.of(workingDir, "modules", namespace + ".jar");
-		for (var i = 0; i < 10; ++i) {
-			try {
-				if (Files.deleteIfExists(moduleDst))
-					break;
-			} catch (Exception ignored) {
-			}
-			//System.out.println("delete module fail=" + i + ":" + moduleDst);
-			Thread.sleep(100);
+		var moduleDstBackup = Path.of(workingDir, "modules", namespace + ".jar.backup").toFile();
+		if (moduleDst.toFile().renameTo(moduleDstBackup)) {
+			txn.whileRollback(() -> {
+				//noinspection ResultOfMethodCallIgnored
+				moduleDstBackup.renameTo(moduleDst.toFile());
+			});
+			txn.whileCommit(() -> Files.deleteIfExists(moduleDstBackup.toPath()));
 		}
 		var moduleDstFile = moduleDst.toFile();
 		if (!moduleSrc.renameTo(moduleDstFile))
@@ -487,8 +563,7 @@ public class HotManager extends ClassLoader {
 		var ready = Path.of(distributeDir, "ready");
 		Task.getScheduledThreadPool().scheduleAtFixedRate(() -> {
 			try {
-				if (Files.exists(ready)) {
-					installReadies();
+				if (Files.exists(ready) && null != installReadies()) {
 					Files.deleteIfExists(ready);
 				}
 			} catch (Throwable ex) {
@@ -609,7 +684,7 @@ public class HotManager extends ClassLoader {
 		return super.findClass(className);
 	}
 
-	public List<HotModule> installReadies() throws Exception {
+	public ArrayList<HotModule> installReadies() throws Exception {
 		var foundJars = new HashSet<String>();
 		var readies = new HashSet<String>();
 		loadExistDistributes(foundJars, readies);
